Java docs implementation
