+++ src/main/java/org/digitalnao/jared/trujillo/exceptions/JsonHandlerException.java
@@ -1,10 +1,20 @@
 package org.digitalnao.jared.trujillo.exceptions;
 
+/**
+ * Unchecked exception for JSON read/parse/mapping failures in this library.
+ */
 public class JsonHandlerException extends RuntimeException {
-    public JsonHandlerException(String message) {
-        super(message);
-    }
-    public JsonHandlerException(String message, Throwable cause) {
-        super(message, cause);
-    }
+    /**
+     * Creates an exception with the given message.
+     * @param message description of the error
+     */
+    public JsonHandlerException(String message) { super(message); }
+
+    /**
+     * Creates an exception with the given message and cause.
+     * @param message description of the error
+     * @param cause underlying cause (may be {@code null})
+     */
+    public JsonHandlerException(String message, Throwable cause) { super(message, cause); }
 }
+
